+++ src/components/Тask.js
@@ -80,4 +80,10 @@ export default class Task extends Component {
   getTemplate() {
     return createTaskTemplate(this._task);
   }
+
+  setEditButtonClickHandler(handler) {
+    this.getElement().querySelector(`.card__btn--edit`)
+      .addEventListener(`click`, handler);
+  }
+
 }
